requirements.txt
aiobotocore
aiohttp
botocore
lxml
mock
Pillow
pytest
pytest-asyncio
tweepy
waybackpy
